fix(security): resolve integer multiplication overflow in paged attention and moe kernels

+++ csrc/turing_cpu_moe_kernel.hpp
@@ -26,8 +26,8 @@ inline void parallel_cpu_moe_gemv(
     // Parallelize across batch items
     pool.parallel_for(0, batch_size, [&](size_t /*worker_id*/, size_t start_b, size_t end_b) {
         for (size_t b = start_b; b < end_b; ++b) {
-            const float* in_vec = input + (b * in_features);
-            float* out_vec = output + (b * out_features);
+            const float* in_vec = input + (b * static_cast<size_t>(in_features));
+            float* out_vec = output + (b * static_cast<size_t>(out_features));
 
             // Initialize output vector to zero
             for (int oc = 0; oc < out_features; ++oc) {
@@ -36,13 +36,13 @@ inline void parallel_cpu_moe_gemv(
 
             // Accumulate contributions from top_k routed experts
             for (int k = 0; k < top_k; ++k) {
-                int exp_idx = expert_indices[b * top_k + k];
-                float gate_w = routing_weights[b * top_k + k];
+                int exp_idx = expert_indices[b * static_cast<size_t>(top_k) + k];
+                float gate_w = routing_weights[b * static_cast<size_t>(top_k) + k];
 
-                const float* w_exp = expert_weights + (exp_idx * out_features * in_features);
+                const float* w_exp = expert_weights + (static_cast<size_t>(exp_idx) * static_cast<size_t>(out_features) * static_cast<size_t>(in_features));
 
                 for (int oc = 0; oc < out_features; ++oc) {
-                    const float* w_row = w_exp + (oc * in_features);
+                    const float* w_row = w_exp + (static_cast<size_t>(oc) * static_cast<size_t>(in_features));
                     float acc = 0.0f;
                     for (int ic = 0; ic < in_features; ++ic) {
                         acc += in_vec[ic] * w_row[ic];
