+++ csrc/turing_paged_attention.hpp
@@ -20,7 +20,7 @@ private:
 public:
     TuringPagedAttentionEngine(int heads, int h_dim, int b_size, int n_blocks)
         : num_heads(heads), head_dim(h_dim), block_size(b_size), total_blocks(n_blocks), kv_pool_raw(nullptr) {
-        page_bytes = 2 * static_cast<size_t>(num_heads) * block_size * head_dim * sizeof(float);
+        page_bytes = 2 * static_cast<size_t>(num_heads) * static_cast<size_t>(block_size) * static_cast<size_t>(head_dim) * sizeof(float);
         size_t total_bytes = static_cast<size_t>(total_blocks) * page_bytes;
         kv_pool_raw = static_cast<uint8_t*>(aligned_alloc_64(total_bytes));
         std::memset(kv_pool_raw, 0, total_bytes);
@@ -34,15 +34,15 @@ public:
     }
 
     float* get_k_block(int physical_block_idx, int head_idx) {
-        uint8_t* page_base = kv_pool_raw + (physical_block_idx * page_bytes);
+        uint8_t* page_base = kv_pool_raw + (static_cast<size_t>(physical_block_idx) * page_bytes);
         float* k_base = reinterpret_cast<float*>(page_base);
-        return k_base + (head_idx * block_size * head_dim);
+        return k_base + (static_cast<size_t>(head_idx) * static_cast<size_t>(block_size) * static_cast<size_t>(head_dim));
     }
 
     float* get_v_block(int physical_block_idx, int head_idx) {
-        uint8_t* page_base = kv_pool_raw + (physical_block_idx * page_bytes);
+        uint8_t* page_base = kv_pool_raw + (static_cast<size_t>(physical_block_idx) * page_bytes);
         float* v_base = reinterpret_cast<float*>(page_base + (page_bytes / 2));
-        return v_base + (head_idx * block_size * head_dim);
+        return v_base + (static_cast<size_t>(head_idx) * static_cast<size_t>(block_size) * static_cast<size_t>(head_dim));
     }
 
     void forward_selective_attention(
@@ -52,14 +52,14 @@ public:
         uint32_t active_page_mask,      // Bitmask for selective page skipping
         float* output_context           // [num_heads, head_dim]
     ) {
-        std::memset(output_context, 0, num_heads * head_dim * sizeof(float));
+        std::memset(output_context, 0, static_cast<size_t>(num_heads) * static_cast<size_t>(head_dim) * sizeof(float));
         float sm_scale = 1.0f / std::sqrt(static_cast<float>(head_dim));
 
         for (int h = 0; h < num_heads; ++h) {
-            const float* q_head = query + (h * head_dim);
-            float* out_head = output_context + (h * head_dim);
+            const float* q_head = query + (static_cast<size_t>(h) * static_cast<size_t>(head_dim));
+            float* out_head = output_context + (static_cast<size_t>(h) * static_cast<size_t>(head_dim));
 
-            std::vector<float> scores(num_logical_pages * block_size, -1e9f);
+            std::vector<float> scores(static_cast<size_t>(num_logical_pages) * static_cast<size_t>(block_size), -1e9f);
             float max_score = -1e9f;
 
             // Phase 1: Sparse QK^T
@@ -71,13 +71,13 @@ public:
                 const float* k_block = get_k_block(pb, h);
 
                 for (int t = 0; t < block_size; ++t) {
-                    const float* k_token = k_block + (t * head_dim);
+                    const float* k_token = k_block + (static_cast<size_t>(t) * static_cast<size_t>(head_dim));
                     float dot = 0.0f;
                     for (int d = 0; d < head_dim; ++d) {
                         dot += q_head[d] * k_token[d];
                     }
                     float score = dot * sm_scale;
-                    scores[lp * block_size + t] = score;
+                    scores[static_cast<size_t>(lp) * static_cast<size_t>(block_size) + static_cast<size_t>(t)] = score;
                     if (score > max_score) {
                         max_score = score;
                     }
@@ -89,9 +89,9 @@ public:
             for (int lp = 0; lp < num_logical_pages; ++lp) {
                 if (!(active_page_mask & (1U << lp))) continue;
                 for (int t = 0; t < block_size; ++t) {
-                    float s = scores[lp * block_size + t];
+                    float s = scores[static_cast<size_t>(lp) * static_cast<size_t>(block_size) + static_cast<size_t>(t)];
                     float exp_val = std::exp(s - max_score);
-                    scores[lp * block_size + t] = exp_val;
+                    scores[static_cast<size_t>(lp) * static_cast<size_t>(block_size) + static_cast<size_t>(t)] = exp_val;
                     sum_exp += exp_val;
                 }
             }
@@ -104,8 +104,8 @@ public:
                 const float* v_block = get_v_block(pb, h);
 
                 for (int t = 0; t < block_size; ++t) {
-                    float weight = scores[lp * block_size + t] * inv_sum;
-                    const float* v_token = v_block + (t * head_dim);
+                    float weight = scores[static_cast<size_t>(lp) * static_cast<size_t>(block_size) + static_cast<size_t>(t)] * inv_sum;
+                    const float* v_token = v_block + (static_cast<size_t>(t) * static_cast<size_t>(head_dim));
                     for (int d = 0; d < head_dim; ++d) {
                         out_head[d] += weight * v_token[d];
                     }
